Move icons into separate folder
