+++ src/renderer/components/icons/CancelIcon.tsx
