+++ src/renderer/components/icons/CanceledIcon.tsx
