+++ src/renderer/components/icons/CompletedIcon.tsx
