+++ src/renderer/components/icons/ConnectIcon.tsx
