+++ src/renderer/components/icons/DeleteIcon.tsx
