+++ src/renderer/components/icons/FailedIcon.tsx
