+++ src/renderer/components/icons/GreenRunIcon.tsx
