+++ src/renderer/components/icons/LoadingIcon.tsx
