+++ src/renderer/components/icons/LogsIcon.tsx
