+++ src/renderer/components/icons/RegistrationIcon.tsx
